pyproject.toml
[tool.poetry]
name = "newsboard"
version = "0.1.0"
description = ""
authors = ["rostislav <rostmoguchiy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.8"
djangorestframework = "^3.11.0"
psycopg2 = "^2.8.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
